Fixing bugs and improving speed
Function createHyperlinksForCitationsAPA was not properly generating the hyperlinks when citations were merged. It turns out that the order of the citations (in the visible text) does not correspond to the order in the citation data (hidden information of the field).

Function createHyperlinksForCitationsIEEE was slow in long documents due to the unnecessary full iterations to locate the citation numbers; it is faster with the use of regular expressions. Besides, it now tells the user what citations are orphan.