Bug when editors are used instead of authors
Function GAUG_createHyperlinksForCitationsAPA had a problem when citations use editors instead of authors (e.g. a book of proceedings). Those citations were not linked and tagged as orphan.
The problem was solved by including "(Ed.)" and "(Eds.)" to the regular expression and including the editors when authors were not found in the citation.