New flag to include square brackets in hyperlinks
There is a new configuration flag that allows to include the square brackets in the hyperlinks. It is not possible to remove the hyperlinks with square brackets and the argument `"RemoveHyperlinks"` or the wrapper `GAUG_removeHyperlinks` because Microsoft Word DOES NOT handle the square brackets in the hyperlinks very well.

Some typos removed.